Moved check for duplicate OnPlayFinishedEvents
The test was done after the call to the active modifier, so this was called twice. This can lead to unexpected behaviour of the modifiers. In fact the check can be moved to the Mp3Notify class. That way we also don't have to create random track numbers if the next button is pressed.